+++ src/main/java/com/grupo3/ProyectoLBD/repository/EncabezadoFacturaViewRepository.java
@@ -8,8 +8,8 @@ import org.springframework.stereotype.Repository;
 import com.grupo3.ProyectoLBD.model.EncabezadoFacturaView;
 
 @Repository
-public interface EncabezadoFacturaViewRepository extends JpaRepository<EncabezadoFacturaView, Integer> {
+public interface EncabezadoFacturaViewRepository extends JpaRepository<EncabezadoFacturaView, Long> {
     List<EncabezadoFacturaView> findByEstadoNot(String estado);
     List<EncabezadoFacturaView> findByCedulaApoderado(Long cedula);
-    List<EncabezadoFacturaView> findByIdFactura(Integer idFactura);
+    List<EncabezadoFacturaView> findByIdFactura(Long idFactura);
 }
